requirements.txt
# Core web framework
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=1.5.0
tabulate==0.9.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LangChain core and community packages
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.50
langchain-google-genai>=0.0.5
langchain-groq>=0.0.3
langchain-core>=0.1.0

# Environment management
python-dotenv>=1.0.0

# HuggingFace support (for free models)
transformers>=4.30.0
torch>=2.0.0

# Fine-tuning support
peft>=0.5.0
bitsandbytes>=0.41.0
accelerate>=0.20.0
datasets>=2.12.0
scikit-learn>=1.3.0
wandb>=0.15.0

# Advanced RAG support
scikit-learn>=1.3.0

# Authentication system
sqlite3

# Enhanced features
numpy>=1.24.0
pillow>=9.5.0
reportlab>=4.0.0
fpdf2>=2.7.0

# Standard library packages (included with Python, listed for reference)
# os - built-in
# io - built-in
# sys - built-in
# re - built-in
# base64 - built-in
# json - built-in
# datetime - built-in